requirements.txt
cachetools==3.1.1
certifi==2023.5.7
charset-normalizer==3.1.0
contourpy==1.0.7
cycler==0.11.0
fonttools==4.39.4
google-api-core==2.11.0
google-api-python-client==1.7.2
google-auth==1.8.0
google-auth-httplib2==0.0.3
google-auth-oauthlib==0.4.1
googleapis-common-protos==1.59.0
httplib2==0.22.0
idna==3.4
kiwisolver==1.4.4
matplotlib==3.7.1
nba-api==1.2
numpy==1.24.3
oauthlib==3.2.2
packaging==23.1
pandas==2.0.1
Pillow==9.5.0
protobuf==4.23.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
pyparsing==3.0.9
python-dateutil==2.8.2
pytz==2023.3
requests==2.31.0
requests-oauthlib==1.3.1
rsa==4.0
six==1.16.0
tzdata==2023.3
uritemplate==3.0.1
urllib3==1.26.16
